+++ src/blockfile/LegacyBlockFile.cpp
@@ -37,7 +37,7 @@
 #include "sndfile.h"
 
 void ComputeLegacySummaryInfo(const wxFileName &fileName,
-                              int summaryLen,
+                              size_t summaryLen,
                               sampleFormat format,
                               SummaryInfo *info,
                               bool noRMS,bool Silent,
